Split site Header and Footer into components This is to prep for making them more complex
